fix int casts

+++ src/systems/debugSystem.h
@@ -18,7 +18,7 @@ void debugSystem(SDL_Renderer* renderer, entt::registry &registry) {
     for (auto entity : cameraView) {
       camera &c = cameraView.get<camera>(entity);
 
-      SDL_Rect dr = { (int)p.x + cr.x - c.pos.x, (int)p.y + cr.y - c.pos.y, (int)cr.w, (int)cr.h };
+      SDL_Rect dr = { (int)p.x + (int)cr.x - (int)c.pos.x, (int)p.y + (int)cr.y - (int)c.pos.y, (int)cr.w, (int)cr.h };
 
       SDL_SetRenderDrawColor(renderer, 0, 255, 0, 255);
       SDL_RenderDrawRect(renderer, &dr);
